+++ flan/tests/utils.py
@@ -2,7 +2,7 @@
 import subprocess
 import re
 import json
-from flan.flan import SUPPORTED_FIELDS, MONTHS, DEFAULT_FORMAT
+from flan.flan import SUPPORTED_FIELDS, DEFAULT_FORMAT
 import shlex
 from dateutil import parser as dtparser
 import os